Adding more  command line parameters to oscars_te: --oscars and --sflow.
This commit adds two extra parameters to the command line of the flask
application: --oscars and --sflow. They are useful to specify the oscars db
and sflow-rt URL at execution time.
This commit also changes the pointer to the OF monitor instance
in the app.config dictionary. Now the instance of the OF monitor is
accessible through the key "controller_instance" of that dictionary.